simplified config.yml. Fixed write to excel such that if file exists then it creates that file, otherwise it writes a new sheet to the existing one

+++ src/ir_pell_accepts/io_utils.py
@@ -178,14 +178,17 @@ def output_results(df: pd.DataFrame, file_path: Path, append_today: bool = True,
     ext = validate_extension(outfile.suffix)
 
     if ext == ".xlsx":
-        with pd.ExcelWriter(
-            outfile,
-            engine="openpyxl",
-            mode="a",
-            if_sheet_exists="replace"
-        ) as writer:
-        
-            df.to_excel(writer, sheet_name="Python Output", index=False)
+        if outfile.exists():
+            with pd.ExcelWriter(
+                outfile,
+                engine="openpyxl",
+                mode="a",
+                if_sheet_exists="replace"
+            ) as writer:
+            
+                df.to_excel(writer, sheet_name="Python Output", index=False)
+        else:
+            df.to_excel(outfile, index=False)
     if ext == ".csv":
         df.to_csv(outfile, index=False)
     if ext == ".txt":
